Move InvalidCommandException to exceptions package
