+++ src/main/java/exceptions/InvalidCommandException.java
@@ -1,4 +1,4 @@
-package commands;
+package exceptions;
 
 public class InvalidCommandException extends Exception {
     public InvalidCommandException(String message) {
